pyproject.toml
[tool.poetry]
name = "sygalin_base_python"
version = "0.1.0"
description = "The base app for future SYGALIN SAS python projects"
authors = ["Harris TIWA <tiwa_m@sygalin.com>"] 
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,< 4.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
asyncpg = "^0.30.0"
alembic = "^1.15.2"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = {extras = ["asyncio"], version = "^5.0.3"}
arq = "^0.25.0"
httpx = "^0.28.1"
cryptography = "^42.0.5"
structlog = "^24.1.0"
pendulum = "^3.0.0"
mangum = "^0.17.0"
pydantic = "^2.6.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.22"
email-validator = "^2.1.0"
